@@ -1,4 +1,4 @@
-// swift-tools-version: 5.9
+// swift-tools-version: 6.0
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -34,9 +34,9 @@ let package = Package(
                 .product(name: "DequeModule", package: "swift-collections")
             ],
             swiftSettings: [
-                .enableExperimentalFeature("StrictConcurrency=complete"),
                 .enableUpcomingFeature("FullTypedThrows"),
-                .enableExperimentalFeature("IsolatedAny")
+                .enableExperimentalFeature("IsolatedAny"),
+                .swiftLanguageVersion(.v6)
             ]
         ),
         .testTarget(
