@@ -20,31 +20,27 @@ public extension Combine.Future {
     @inlinable
     final var compatValue: Output {
         get async throws(Failure) {
-            if #available(iOS 15.0, tvOS 15.0, macCatalyst 15.0, watchOS 8.0, macOS 12.0, *) {
-                return try await value
-            } else {
-                let result: Result<Output,Failure> = await withCheckedContinuation { continuation in
-                    self.subscribe(AnySubscriber(
-                        receiveSubscription: {
-                            $0.request(.max(1))
-                        },
-                        receiveValue: { (value: sending Output) in
-                            continuation.resume(returning: .success(value))
-                            return .none
-                        },
-                        receiveCompletion: {
-                            if case let .failure(error) = $0 {
-                                continuation.resume(returning: .failure(error))
-                            }
+            let result: Result<Output,Failure> = await withCheckedContinuation { continuation in
+                self.subscribe(AnySubscriber(
+                    receiveSubscription: {
+                        $0.request(.max(1))
+                    },
+                    receiveValue: { (value: sending Output) in
+                        continuation.resume(returning: .success(value))
+                        return .none
+                    },
+                    receiveCompletion: {
+                        if case let .failure(error) = $0 {
+                            continuation.resume(returning: .failure(error))
                         }
-                    ))
-                }
-                switch result {
-                case .success(let success):
-                    return success
-                case .failure(let failure):
-                    throw failure
-                }
+                    }
+                ))
+            }
+            switch result {
+            case .success(let success):
+                return success
+            case .failure(let failure):
+                throw failure
             }
         }
     }
