@@ -88,7 +88,7 @@ func wrapToResult<T,Failure:Error>(_ block: () throws(Failure) -> T) -> Result<T
     }
 }
 
-@available(macOS 9999, *)
+@available(macOS 15.0, iOS 18.0, watchOS 11.0, tvOS 18.0, visionOS 2.0, *)
 @inline(__always)
 @usableFromInline
 internal
