@@ -22,7 +22,7 @@ class AsyncSequencePublisherTests: XCTestCase {
             source.forEach{ continuation.yield($0) }
             continuation.finish()
         }
-        let cancellable = AsyncSequencePublisher(base: stream)
+        let cancellable = AsyncSequencePublisher(legacy: stream)
             .catch{ _ in
                 XCTFail()
                 return Empty<Int, Never>()
@@ -59,7 +59,7 @@ class AsyncSequencePublisherTests: XCTestCase {
             }
             return value
         }
-        let pub = AsyncSequencePublisher(base: asyncSequence)
+        let pub = AsyncSequencePublisher(legacy: asyncSequence)
             .handleEvents(
                 receiveCancel: { expect.fulfill() }
             )
@@ -88,7 +88,7 @@ class AsyncSequencePublisherTests: XCTestCase {
             }
             return value
         }
-        let cancellable = AsyncSequencePublisher(base: asyncSequence)
+        let cancellable = AsyncSequencePublisher(legacy: asyncSequence)
             .mapError{
                 $0 as! CancellationError
             }
